+++ capstone-rs/src/arch/sparc.rs
@@ -24,7 +24,7 @@ pub enum SparcOperand {
     Reg(RegId),
 
     /// Immediate
-    Imm(i32),
+    Imm(i64),
 
     /// Memory
     Mem(SparcOpMem),