+++ capstone-rs/src/instruction.rs
@@ -263,7 +263,7 @@ impl_SliceIterator_wrapper!(
 
 impl<'a> InsnDetail<'a> {
     /// Returns the implicit read registers
-    pub fn regs_read(&self) -> RegsIter<u8> {
+    pub fn regs_read(&self) -> RegsIter<RegIdInt> {
         RegsIter((*self.0).regs_read[..self.regs_read_count() as usize].iter())
     }
 
@@ -273,7 +273,7 @@ impl<'a> InsnDetail<'a> {
     }
 
     /// Returns the implicit write registers
-    pub fn regs_write(&self) -> RegsIter<u8> {
+    pub fn regs_write(&self) -> RegsIter<RegIdInt> {
         RegsIter((*self.0).regs_write[..self.regs_write_count() as usize].iter())
     }